+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](http://semver.org/).
 - `notFound` response.
 - `payloaded` `Prism'` for `Request` `Payload`s.
 - Smart constructors for `GET` and `POST`.
+- `HasBody` instance for `Request`.
 
 ## [0.2] - 2015-4-5
 ### Changed